\name{countries_lib-package}
\alias{countries_lib-package}
\alias{countries_lib}
\docType{package}
\title{
\packageTitle{countries_lib}
}
\description{
\packageDescription{countries_lib}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{countries_lib}
\packageIndices{countries_lib}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{countries_lib}

Maintainer: \packageMaintainer{countries_lib}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
